# repo/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospitality Operations Management System API
  description: Backend API for managing events, resources, reservations, recipes, inventory, entitlements, and operational workflows.
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login or POST /auth/refresh.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error
        - message
...
    Entitlement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type_id:
          type: string
          format: uuid
        holder_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [active, redeemed, expired, revoked]
        issued_at:
          type: string
          format: date-time
        redeemed_at:
          type: string
          format: date-time
          nullable: true
        idempotency_key:
          type: string
          nullable: true
...
paths:
...
  /entitlements:
    get:
      tags: [Entitlements]
      summary: List entitlements
      parameters:
        - name: holder_id
          in: query
          schema:
            type: string
            format: uuid
        - name: event_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, redeemed, expired, revoked]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated entitlements
    post:
      tags: [Entitlements]
      summary: Issue an entitlement (manual)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type_id, holder_id]
              properties:
                type_id:
                  type: string
                  format: uuid
                holder_id:
                  type: string
                  format: uuid
                event_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Entitlement issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entitlement'
...
  /entitlements/{id}/redeem:
    post:
      tags: [Entitlements]
      summary: Redeem an entitlement (idempotent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idempotency_key]
              properties:
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Entitlement redeemed (or already redeemed with same key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entitlement'
        '400':
          description: Already redeemed with different key or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...